pyproject.toml
[tool.poetry]
name = "credit_recovery"
version = "0.1.0"
description = ""
authors = ["Marlon Leite <marlonleite@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.2"
djangorestframework = "^3.12.4"
requests = "^2.25.1"
python-decouple = "^3.4"
validate-docbr = "^1.8.2"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
ipython = "^7.23.1"
django-extensions = "^3.1.3"
black = "^21.5b1"
flake8 = "^3.9.2"
pytest = "^6.2.4"
pytest-django = "^4.2.0"
pytest-mock = "^3.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
